doom: fix abs() macro on unsigned angle_t — kills "gray slab" walls
The vendored doomgeneric calls abs() on unsigned angle_t values in
several places, most importantly:

    offsetangle = abs(rw_normalangle - rw_angle1);   // r_segs.c

Real libc int abs(int) takes a signed argument, so passing an
angle_t (unsigned int) implicitly reinterprets large unsigned values
as negative ints and abs returns their positive equivalent — which
is exactly the "shortest angle" behaviour the renderer needs when
the unsigned subtraction wraps.

Our dglibc_compat.h shimmed abs as

    #define abs(x) ((x) < 0 ? -(x) : (x))

For unsigned x, `(x) < 0` is always false, so the macro was a no-op
and the unsigned wrap survived unchanged. The subsequent
`if (offsetangle > ANG90) offsetangle = ANG90;` then mis-clamped a
small wrapped negative angle (e.g. -5° showing up as 0xFFFFFFFB ≈
ANG358) all the way to ANG90, which made

    distangle = ANG90 - offsetangle;
    sineval   = finesine[distangle >> ANGLETOFINESHIFT];
    rw_distance = FixedMul(hyp, sineval);

collapse to ~0 for entire 1-sided wall segs at normal viewing
distance. R_ScaleFromGlobalAngle hit its 64*FRACUNIT clamp, the
per-column adjustment FixedMul(finetangent[angle], rw_distance)
became <1 fixed unit, so

    texturecolumn = (rw_offset - adj) >> FRACBITS

returned the same value for every screen X — i.e. one texture
column repeated across the whole seg, which renders as a uniform
"gray slab" with no apparent Doom texture detail.

Cast to (int) inside the macro so abs(unsigned_angle_t) behaves
the way the renderer expects.

+++ kernel/doom/dglibc_compat.h
@@ -210,9 +210,25 @@ extern double cos(double x);
 extern double atan2(double y, double x);
 extern double pow(double x, double y);
 extern double log(double x);
-/* Integer abs */
+/* Integer abs.
+ *
+ * Cast to (int) first so callers passing an unsigned value (notably
+ * angle_t in r_segs.c / r_main.c) get the same "shortest angle"
+ * behaviour as the real libc int abs(int): for unsigned x>INT_MAX,
+ * (int)x is negative, abs returns its positive equivalent (i.e.
+ * 2^32 - x). Without the cast, (x)<0 is always false on unsigned and
+ * the macro is a no-op, which makes
+ *
+ *     abs(rw_normalangle - rw_angle1)
+ *
+ * mis-clamp to ANG90 whenever the true offset angle is small but the
+ * unsigned subtraction wraps — that produced rw_distance ≈ 0 and
+ * forced R_ScaleFromGlobalAngle into its 64*FRACUNIT clamp, which in
+ * turn made R_RenderSegLoop sample the same texture column for every
+ * screen X (a uniform vertical stripe = "gray slab" walls).
+ */
 #ifndef abs
-#define abs(x)  ((x) < 0 ? -(x) : (x))
+#define abs(x) (((int)(x)) < 0 ? -((int)(x)) : ((int)(x)))
 #endif
 
 /* ------------------------------------------------------------------ */
